projects/vstudio: enthusiastic warnings
The commit removes the three compile warning options from the individual project
files into the zlib.props globals (this should probably be renamed in 1.7).  It
increases the warning level from 4 to All and adds a list of the warnings which
need to be turned off.  This is semi-documentary; the intent is to tell users of
libpng which warnings have been examined and judged non-fixable at present.

The warning about structure padding *IS* fixable, but it is a signficant change
(moving structure members around).  It is currently partially fixed in 1.7, but
only partially.

Signed-off-by: John Bowler <jbowler@acm.org>