added PNG_ARM_NEON switch and enabled building with NEON support
CMAKE_SYSTEM_PROCESSOR is checked to discern whether to build for arm or not, as it is the canonical variable for this.
Behaviour is copied from autoconf; by default PNG_ARM_NEON is set to 'check' and NEON support is checked dynamically at runtime. Other possible values are 'on' and 'off'. The deprecated option 'api' is not used here.