added PNG_ARM_NEON switch and enabled building with NEON support
CMAKE_SYSTEM_PROCESSOR is checked to discern whether to build for arm or not, as it is the canonical variable for this.
Behaviour is copied from autoconf; by default PNG_ARM_NEON is not set. If it is manually set to 'api', 'on' or 'off' the same defines as in using autotools are defined. The deprecated 'check' option is not supported here.