Fix processing of background color when converting to grayscale
When processing transparent pixels (with png_set_strip_alpha) while
converting to grayscale (with png_set_rgb_to_gray), libpng failed to
convert the background color specified to the user to grayscale as
well. This caused the documented pattern of handling bKGD chunks to
fail for this corner case.

Note that in case of conversions in the opposite direction (grayscale
to RGB), the background color was already properly converted to RGB.

This patch fixes this inconsistency, and adds a test program to the
test suite to verify the fix.