CMakeLists.txt: Improve pkg-config generating on Windows
Use the target properties of ZLIB::ZLIB to get the library names for the
pkg-config files that we generate, and unify things with other builds.
We need to use the IMPORT_LOCATION[_$<CONFIG>] for ZLib since it does not
define IMPORT_IMPLIB[_$<CONFIG>] even on Windows DLL builds, and assume
that the $<CONFIG> we use above is RELEASE for both MinSizeRel and
RelWithDebInfo builds.