arm/filter_neon.S: remove obsolete assembler
This file contains hand-coded assembler implementations of the filter
functions for 32-bit Arm platforms.  These are only used when the
compiler doesn't support neon intrinsics (added to GCC 4.3 in 2008) or
is exactly GCC 4.5.4 (released 2012), both of which are sufficiently
unlikely to be true that it's fair to say the assembler is no longer
used.

Instrinics can be optimised by the compiler, and also has the benefit
that the compiler can add branch protection instructions (such as
PAC/BTI on aarch64) to improve resilience against malicious input.

This commit deletes filter_neon.S and removes the now obsolete
preprocessor logic in pngpriv.h.

Closes #505