png_safe_execute setjmp and GCC warning fixes
png_safe_execute called setjmp in a context where the result is
undefined (an assignment statement).  This corrects the code and removes
volatile statements that were introduced previously to quell warnings
from earlier versions of GCC.  Appropriate 'const' qualificationos now
show that no auto/register variable which is in extent when setjmp is
called is modified after initial assignment anywhere in the function; a
superset of the ANSI-C requirements which also aids code maintenance.

Tests: configure:

  CFLAGS="-Wall -Wextra -Wno-maybe-uninitialized" ../configure --enable-werror
  make
  make check

cmake:

  cmake ..
  make
  make test

Signed-off-by: John Bowler <jbowler@acm.org>