CMakeLists: Enable pkg-config on Windows
In libpng, it's hard-coded to exclude pkg-config on Windows unless it's MinGW or Cygwin.
I use pkg-config with Visual Studio 2022 and I have to patch this out for CMake to install the pc file.
I'm not sure what the original intention was here, installing the .pc file does no harm even if you do not use pkg-config, but if someone really needs to disable it, it should be a CMake option, not hard-coded like this. create_symlink() also works on Windows. Out of over 50 libraries I build on Windows, libpng is the only one that have this behavior.

Co-Authored-By: Miloš Komarčević <4973094+kmilos@users.noreply.github.com>